removed extra pojo, move listener in different class
