+++ src/main/java/org/jointheleague/discord_bot_example/Bot.java
@@ -5,7 +5,7 @@ import org.javacord.api.DiscordApiBuilder;
 import org.javacord.api.event.message.MessageCreateEvent;
 import org.javacord.api.listener.message.MessageCreateListener;
 
-public class Bot implements MessageCreateListener {
+public class Bot  {
 	private String token;
 	private String channelName;
 	DiscordApi api;
@@ -19,22 +19,9 @@ public class Bot implements MessageCreateListener {
 		api = new DiscordApiBuilder().setToken(token).login().join();
 		System.out.println("You can invite the bot by using the following url: " + api.createBotInvite());
 		api.getServerTextChannelsByName(channelName).forEach(e -> e.sendMessage("Bot Connected"));
-		api.addMessageCreateListener(this);
-	}
-
-	@Override
-	public void onMessageCreate(MessageCreateEvent event) {
-		event.getServerTextChannel().ifPresent(e -> {
-			if (e.getName().equals(channelName)) {
-				if (event.getMessageContent().equalsIgnoreCase("!ping")) {
-					event.getChannel().sendMessage("Pong!");
-				} else if (event.getMessageContent().equalsIgnoreCase("!bing")) {
-					event.getChannel().sendMessage("Bong!");
-				}
-			}
-
-		});
-
+		api.addMessageCreateListener(new MessageListener(channelName));
 	}
 
 }
+
+
