--- src/main/java/org/jointheleague/discord_bot_example/pojos/BotList.java
@@ -1,21 +1,0 @@
-package org.jointheleague.discord_bot_example.pojos;
-
-import java.util.List;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class BotList {
-
-	@SerializedName("bots")
-	@Expose
-	private List<BotInfo> bots = null;
-
-	public List<BotInfo> getBots() {
-		return bots;
-	}
-
-	public void setBots(List<BotInfo> bots) {
-		this.bots = bots;
-	}
-
-}
